perf: match equations: use diteInduction to split
This PR lets the match equation code use `diteInduction` instead of the
`split` tactic. This is much more surgical, and does not require
separate steps for splitting and for rewriting, and in particular avoids
using the simplifier to rewrite the splitted `ite`, thus avoiding
traversing the expression in every match.